add style guide

+++ src/app/controllers/ChefController.js
@@ -3,46 +3,52 @@ const File = require('../models/File');
 
 async function getRecipeImage(recipe, req) {
   const results = await File.findByRecipe(recipe.id);
-  const files = results.rows.map(file => ({
+  const files = results.rows.map((file) => ({
     ...file,
-    src: `${req.protocol}://${req.headers.host}${file.path.replace('public', '')}`
+    src: `${req.protocol}://${req.headers.host}${file.path.replace(
+      'public',
+      ''
+    )}`,
   }));
-  
+
   return {
     ...recipe,
-    files
-  }
-};
+    files,
+  };
+}
 
 async function getChefImage(chef, req) {
   const results = await File.findOne(chef.file_id);
-  const files = results.rows.map(file => ({
+  const files = results.rows.map((file) => ({
     ...file,
-    src: `${req.protocol}://${req.headers.host}${file.path.replace('public', '')}`
+    src: `${req.protocol}://${req.headers.host}${file.path.replace(
+      'public',
+      ''
+    )}`,
   }));
-  
+
   return {
     ...chef,
-    files
-  }
-};
+    files,
+  };
+}
 
 module.exports = {
   async index(req, res) {
     try {
-      const isAdmin = req.session.isAdmin;
+      const { isAdmin } = req.session;
 
       const results = await Chef.all();
       let chefs = results.rows;
-  
-      const filesPromise = chefs.map(chef => getChefImage(chef, req));
+
+      const filesPromise = chefs.map((chef) => getChefImage(chef, req));
       chefs = await Promise.all(filesPromise);
-  
-      return res.render('chef/index', { chefs, isAdmin });
-    } catch (err) {
-      console.error('ChefsController index', err);
 
       return res.render('chef/index', { chefs, isAdmin });
+    } catch (err) {
+      console.error(err);
+
+      return res.render('chef/index');
     }
   },
 
@@ -51,40 +57,42 @@ module.exports = {
   },
 
   async show(req, res) {
-    const isAdmin = req.session.isAdmin;
-    
+    const { isAdmin } = req.session;
+
     const chef = await getChefImage(req.chef, req);
 
     const results = await Chef.findRecipeByChef(chef.id);
-    const recipesFilesPromise = results.rows.map(recipe => getRecipeImage(recipe, req));
+    const recipesFilesPromise = results.rows.map((recipe) =>
+      getRecipeImage(recipe, req)
+    );
     const recipes = await Promise.all(recipesFilesPromise);
 
     return res.render('chef/show', { chef, recipes, isAdmin });
   },
-  
+
   async post(req, res) {
     try {
-      const { name } = req.body
+      const { name } = req.body;
 
-      let results = await  File.create(...req.files);
-      const fileId = results.rows[0].id
-      
-      results = await Chef.create(name, fileId)
-      const chefId = results.rows[0].id
+      let results = await File.create(...req.files);
+      const fileId = results.rows[0].id;
+
+      results = await Chef.create(name, fileId);
+      const chefId = results.rows[0].id;
 
       results = await Chef.findOne(chefId);
       const chef = await getChefImage(results.rows[0], req);
 
       return res.render('chef/show', {
         chef,
-        success: `O chef ${chef.name} foi criado com sucesso.`
+        success: `O chef ${chef.name} foi criado com sucesso.`,
       });
     } catch (err) {
       console.error('ChefController post', err);
 
       return res.render('chef/create', {
         chef: req.body,
-        error: 'Erro inesperado, tente novamente.'
+        error: 'Erro inesperado, tente novamente.',
       });
     }
   },
@@ -97,7 +105,7 @@ module.exports = {
   async put(req, res) {
     try {
       const removedImges = req.body.removed_images;
-      let { id: chefId, file_id: fileId } = req.chef;
+      const { id: chefId, file_id: fileId } = req.chef;
 
       let data = { ...req.body, fileId };
 
@@ -110,49 +118,50 @@ module.exports = {
 
       await Chef.update(data);
 
-      if (removedImges) 
-        await File.delete(removedImges);
-      
-      results = await Chef.findOne(chefId);
+      if (removedImges) await File.delete(removedImges);
+
+      let results = await Chef.findOne(chefId);
       const chef = await getChefImage(results.rows[0], req);
 
       results = await Chef.findRecipeByChef(chefId);
-      const recipesFilesPromise = results.rows.map(recipe => getRecipeImage(recipe, req));
+      const recipesFilesPromise = results.rows.map((recipe) =>
+        getRecipeImage(recipe, req)
+      );
       const recipes = await Promise.all(recipesFilesPromise);
 
       return res.render('chef/show', {
         chef,
         recipes,
-        success: `O chef ${chef.name} foi atualizado com sucesso.`
+        success: `O chef ${chef.name} foi atualizado com sucesso.`,
       });
     } catch (err) {
       console.error('ChefController put', err);
 
       return res.render('chef/edit', {
         error: 'Erro inesperado, tente novamanete.',
-        chef: req.body
-      })
+        chef: req.body,
+      });
     }
   },
 
   async delete(req, res) {
     try {
-      const chef = req.chef;
+      let { chef } = req;
 
       const results = await Chef.findRecipeByChef(chef.id);
-      const recipes =  results.rows;
-  
-      if (recipes.length == 0) {
+      const recipes = results.rows;
+
+      if (recipes.length === 0) {
         await Chef.delete(chef.id);
         await File.delete(chef.file_id);
 
-        const results = await Chef.all();
-        const filesPromise = results.rows.map(chef => getChefImage(chef, req));
-        const chefs = await Promise.all(filesPromise);
-  
+        const chefs = await Chef.all();
+        const filesPromise = chefs.map((chef) => getChefImage(chef, req));
+        const sereializedChefs = await Promise.all(filesPromise);
+
         return res.render('chef/index', {
-          chefs,
-          success: `o chef ${chef.name} deletado com sucesso.`
+          chefs: sereializedChefs,
+          success: `o chef ${chef.name} deletado com sucesso.`,
         });
       }
 
@@ -160,15 +169,15 @@ module.exports = {
 
       return res.render('chef/edit', {
         chef,
-        error: 'Esse chef não pode ser deletado, pois possui pelo menos receita cadastrada!'
-      })
+        error:
+          'Esse chef não pode ser deletado, pois possui pelo menos receita cadastrada!',
+      });
     } catch (err) {
       console.error('ChefController delete', err);
 
       return res.render('chef/edit', {
-        chef,
-        error: 'Erro inesperado, tente novamente.'
-      })
+        error: 'Erro inesperado, tente novamente.',
+      });
     }
   },
 };
